refactor(General): :recycle: Rename prometheus an rest extentions in .api
