+++ src/pages/Addresses/views/Addresses.tsx
@@ -1,14 +1,14 @@
 import { Card, Grid, GridItem } from '@patternfly/react-core';
 import { useQuery } from '@tanstack/react-query';
 
-import { RESTApi } from '@API/REST';
+import { PrometheusApi } from '@API/Prometheus.api';
+import { RESTApi } from '@API/REST.api';
 import { DEFAULT_TABLE_PAGE_SIZE } from '@config/config';
 import { isPrometheusActive } from '@config/Prometheus.config';
 import SkTable from '@core/components/SkTable';
 import SkTitle from '@core/components/SkTitle';
 import TransitionPage from '@core/components/TransitionPages/Slide';
 import LoadingPage from '@pages/shared/Loading';
-import { PrometheusApi } from 'API/Prometheus';
 
 import { addressesColumns, addressesColumnsWithFlowPairsCounters, customAddressCells } from '../Addresses.constants';
 import { AddressesLabels } from '../Addresses.enum';
