+++ src/pages/ProcessGroups/views/ProcessGroup.tsx
@@ -4,8 +4,8 @@ import { Grid, GridItem } from '@patternfly/react-core';
 import { useQuery } from '@tanstack/react-query';
 import { useParams } from 'react-router-dom';
 
+import { RESTApi } from '@API/REST.api';
 import { isPrometheusActive } from '@config/Prometheus.config';
-import { RESTApi } from '@API/REST';
 import { getTestsIds } from '@config/testIds.config';
 import SkTable from '@core/components/SkTable';
 import SkTitle from '@core/components/SkTitle';
