+++ src/pages/Processes/views/ProcessPairs.tsx
@@ -5,7 +5,7 @@ import { LongArrowAltLeftIcon, LongArrowAltRightIcon } from '@patternfly/react-i
 import { useQuery } from '@tanstack/react-query';
 import { useParams, useSearchParams } from 'react-router-dom';
 
-import { RESTApi } from '@API/REST';
+import { RESTApi } from '@API/REST.api';
 import { AvailableProtocols, SortDirection, TcpStatus } from '@API/REST.enum';
 import { DEFAULT_TABLE_PAGE_SIZE, UPDATE_INTERVAL } from '@config/config';
 import LinkCell from '@core/components/LinkCell';
