+++ src/pages/shared/Metrics/services/index.ts
@@ -1,5 +1,4 @@
-import { PrometheusApi } from '@API/Prometheus';
-import { defaultTimeInterval, timeIntervalMap } from '@config/Prometheus.config';
+import { PrometheusApi } from '@API/Prometheus.api';
 import { PrometheusApiResult } from '@API/Prometheus.interfaces';
 import {
   extractPrometheusLabels,
@@ -7,6 +6,7 @@ import {
   extractPrometheusValuesAndLabels
 } from '@API/Prometheus.utils';
 import { AvailableProtocols } from '@API/REST.enum';
+import { defaultTimeInterval, timeIntervalMap } from '@config/Prometheus.config';
 import { formatToDecimalPlacesIfCents } from '@core/utils/formatToDecimalPlacesIfCents';
 import { getCurrentAndPastTimestamps } from '@core/utils/getCurrentAndPastTimestamps';
